+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # Change Log
 
+## 1.1.0 (Not Releases Yet)
+
+* The `CorsExceptionHandler` class was deprecated and no more to used in the `exceptions.php` file.
+
 ## 1.0.1 (2021-05-07)
 
-Fix matching paths and add test cases.
+* Fix matching paths and add test cases.
 
 ## 1.0.0 (2021-04-26)
 
-Implements from [fruitcake/laravel-cors](https://github.com/fruitcake/laravel-cors) for Hyperf.
+* Implements from [fruitcake/laravel-cors](https://github.com/fruitcake/laravel-cors) for Hyperf.